# tp2/src/openapi.yaml
openapi: 3.1.0
info:
  title: API de depósito
  version: "1.0.0"
  description: |
    Categorías y productos, lugares (depósito, pasillo, fila, estante, bin)
    y stock solo en bins. Un ingreso asigna la cantidad escaneada a uno o más
    bins y la suma una sola vez al completarse. Un pedido se pickea desde bins
    hacia un cartón y se embala cuando cada línea está completa.
servers:
  - url: http://localhost:3000
...
paths:
...
  /warehouses:
    get:
      tags: [Depósitos]
      operationId: listWarehouses
      summary: Listar depósitos
      responses:
        "200":
          description: Depósitos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      tags: [Depósitos]
      operationId: createWarehouse
      summary: Crear un depósito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseWrite"
      responses:
        "201":
          description: Depósito creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /warehouses/{warehouseId}:
    parameters:
      - $ref: "#/components/parameters/warehouseId"
    get:
      tags: [Depósitos]
      operationId: getWarehouse
      summary: Leer un depósito
      responses:
        "200":
          description: Depósito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Depósitos]
      operationId: updateWarehouse
      summary: Cambiar código o nombre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehousePatch"
      responses:
        "200":
          description: Depósito actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Depósitos]
      operationId: deleteWarehouse
      summary: Borrar un depósito sin pasillos
      responses:
        "204":
          description: Borrado
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /warehouses/{warehouseId}/aisles:
    parameters:
      - $ref: "#/components/parameters/warehouseId"
    get:
      tags: [Pasillos]
      operationId: listAisles
      summary: Listar pasillos de un depósito
      responses:
        "200":
          description: Pasillos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Aisle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Pasillos]
      operationId: createAisle
      summary: Crear un pasillo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeWrite"
      responses:
        "201":
          description: Pasillo creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aisle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /aisles/{aisleId}:
    parameters:
      - $ref: "#/components/parameters/aisleId"
    get:
      tags: [Pasillos]
      operationId: getAisle
      summary: Leer un pasillo
      responses:
        "200":
          description: Pasillo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aisle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Pasillos]
      operationId: updateAisle
      summary: Cambiar el código
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeWrite"
      responses:
        "200":
          description: Pasillo actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aisle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Pasillos]
      operationId: deleteAisle
      summary: Borrar un pasillo sin filas
      responses:
        "204":
          description: Borrado
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /aisles/{aisleId}/rows:
    parameters:
      - $ref: "#/components/parameters/aisleId"
    get:
      tags: [Filas]
      operationId: listRows
      summary: Listar filas de un pasillo
      responses:
        "200":
          description: Filas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Row"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Filas]
      operationId: createRow
      summary: Crear una fila
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeWrite"
      responses:
        "201":
          description: Fila creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Row"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /rows/{rowId}:
    parameters:
      - $ref: "#/components/parameters/rowId"
    get:
      tags: [Filas]
      operationId: getRow
      summary: Leer una fila
      responses:
        "200":
          description: Fila
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Row"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Filas]
      operationId: updateRow
      summary: Cambiar el código
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeWrite"
      responses:
        "200":
          description: Fila actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Row"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Filas]
      operationId: deleteRow
      summary: Borrar una fila sin estantes
      responses:
        "204":
          description: Borrada
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /rows/{rowId}/shelves:
    parameters:
      - $ref: "#/components/parameters/rowId"
    get:
      tags: [Estantes]
      operationId: listShelves
      summary: Listar estantes de una fila
      responses:
        "200":
          description: Estantes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shelf"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Estantes]
      operationId: createShelf
      summary: Crear un estante
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeWrite"
      responses:
        "201":
          description: Estante creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shelf"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /shelves/{shelfId}:
    parameters:
      - $ref: "#/components/parameters/shelfId"
    get:
      tags: [Estantes]
      operationId: getShelf
      summary: Leer un estante
      responses:
        "200":
          description: Estante
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shelf"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Estantes]
      operationId: updateShelf
      summary: Cambiar el código
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeWrite"
      responses:
        "200":
          description: Estante actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shelf"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Estantes]
      operationId: deleteShelf
      summary: Borrar un estante sin bins
      responses:
        "204":
          description: Borrado
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /shelves/{shelfId}/bins:
    parameters:
      - $ref: "#/components/parameters/shelfId"
    get:
      tags: [Bins]
      operationId: listBins
      summary: Listar bins de un estante
      responses:
        "200":
          description: Bins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bin"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Bins]
      operationId: createBin
      summary: Crear un bin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeWrite"
      responses:
        "201":
          description: Bin creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bin"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /bins/{binId}:
    parameters:
      - $ref: "#/components/parameters/binId"
    get:
      tags: [Bins]
      operationId: getBin
      summary: Leer un bin
      responses:
        "200":
          description: Bin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bin"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Bins]
      operationId: updateBin
      summary: Cambiar el código
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeWrite"
      responses:
        "200":
          description: Bin actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bin"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Bins]
      operationId: deleteBin
      summary: Borrar un bin sin stock ni movimientos
      responses:
        "204":
          description: Borrado
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /bins/{binId}/stock:
    parameters:
      - $ref: "#/components/parameters/binId"
    get:
      tags: [Bins]
      operationId: getBinStock
      summary: Saldos de un bin
      responses:
        "200":
          description: Saldos del bin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BinStock"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    warehouseId:
      name: warehouseId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    aisleId:
      name: aisleId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    rowId:
      name: rowId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    shelfId:
      name: shelfId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    binId:
      name: binId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  responses:
    BadRequest:
      description: El cuerpo o el id no cumplen el contrato
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No existe ese recurso
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: El estado actual no permite la operación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Warehouse:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
    WarehouseWrite:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
    WarehousePatch:
      type: object
      minProperties: 1
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
    CodeWrite:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 1
    Aisle:
      type: object
      required: [id, code, warehouseId]
      properties:
        id:
          type: integer
        code:
          type: string
        warehouseId:
          type: integer
    Row:
      type: object
      required: [id, code, aisleId]
      properties:
        id:
          type: integer
        code:
          type: string
        aisleId:
          type: integer
    Shelf:
      type: object
      required: [id, code, rowId]
      properties:
        id:
          type: integer
        code:
          type: string
        rowId:
          type: integer
    Bin:
      type: object
      required: [id, code, shelfId]
      properties:
        id:
          type: integer
        code:
          type: string
        shelfId:
          type: integer
    BinBalance:
      type: object
      required: [productId, sku, name, quantity]
      properties:
        productId:
          type: integer
        sku:
          type: string
        name:
          type: string
        quantity:
          type: integer
          minimum: 0
    BinStock:
      type: object
      required: [binId, balances]
      properties:
        binId:
          type: integer
        balances:
          type: array
          items:
            $ref: "#/components/schemas/BinBalance"
...